+++ package.json
@@ -8,7 +8,7 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "mocha",
+    "test": "mocha --exit",
     "start": "nodemon server.js"
   },
   "repository": {
@@ -46,6 +46,7 @@
     "karma-opera-launcher": "^1.0.0",
     "karma-safari-launcher": "^1.0.0",
     "mocha": "^5.0.1",
+    "passport-mock": "0.0.3",
     "watchify": "^3.10.0"
   }
 }
