+++ test/dashboardTest.js
@@ -1,8 +1,8 @@
 process.env.NODE_ENV = 'test';
 
 const expect = require('chai').expect;
-const dashboard = require('../src/controllers/dashboard');
-const server = require('../src/server.js');
+const dashboard = require('../controllers/dashboard');
+const server = require('../server.js');
 
 describe('Dashboard Controller', function() {
   it('should render pages successfully', function() {});
