+++ test/devdashClientTest.js
@@ -1,7 +1,7 @@
 process.env.NODE_ENV = 'test';
 
 const expect = require('chai').expect;
-const script = require('../src/public/js/devdash.js');
+const script = require('../public/js/devdash.js');
 
 describe('Client-side JS', function() {
   describe('Dashboard Page', function() {});
