+++ test/serverTest.js
@@ -1,7 +1,7 @@
 process.env.NODE_ENV = 'test';
 
 const expect = require('chai').expect;
-const server = require('../src/server');
+const server = require('../server');
 
 describe('Server', function() {
   it('should start without erroring', function() {});
