+++ util/helpers.js
@@ -2,7 +2,7 @@ const ensureAuthenticated = function(req, res, next) {
   if (req.isAuthenticated()) {
     return next();
   }
-  res.status(401).redirect('/');
+  res.status(401).send('Unauthorized!');
 };
 
 module.exports = {
